pyproject.toml
[tool.poetry]
name        = "carbon-diff"
version     = "0.1.0"
description = ""
authors     = ["Omair <al3dwii@outlook.com>"]
readme      = "README.md"
packages    = [{ include = "carbon_diff" }]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
pandas = "2.2.2"
typer  = "0.12.3"
click  = "8.1.7"        # keep Click 8 for Typer 0.12.x
pyyaml = "6.0.1"
rich   = "13.7.1"
boto3 = "1.34.101"
fastapi = {extras = ["all"], version = "^0.115.12"}
uvicorn = {version = "0.29.0", extras = ["standard"]}
pyarrow = "^15.0"
google-cloud-bigquery = "^3.20.0"
sqlmodel = "0.0.16"
aiosqlite = "^0.21.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
ruff   = "^0.4.4"

[tool.poetry.scripts]
carbon-diff = "carbon_diff.cli:app"
carbon-record = "carbon_diff.ledger.record:app"

[build-system]
requires      = ["poetry-core>=1.7.0,<2.0.0"]
build-backend = "poetry.core.masonry.api"
